+++ noodles-bam/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Changed
 
+  * bam/bai/async/reader: Disallow duplicate bin IDs.
+
   * bam/lazy/record: Change `lazy::Record::flags` to be infallible.
 
 ## 0.44.0 - 2023-08-31